Remove custom `PageSizeSelect` and `PaginationButtons` components
